refactor: move buttons into separate folder
it's confusing to have items that don't use the "<zoa-input>" format in the inputs folder, so they're now in a separate "buttons" folder
