+++ src/components/index.js
@@ -1,5 +1,6 @@
-export { default as ZoaButton } from './inputs/button/Button.vue';
-export { default as ZoaToggleButton } from './inputs/button/ToggleButton.vue';
+export { default as ZoaButton } from './buttons/button/Button.vue';
+export { default as ZoaToggleButton } from './buttons/togglebutton/ToggleButton.vue';
+
 export { default as ZoaInput } from './inputs/Input.vue';
 
 export { default as ZoaFlash } from './dialogs/flash/Flash.vue';
